fix(ci): prettier api src; clippy lending/governance
- api: prettier --write all src/**/*.ts for format step
- lending: crate-level allow(too_many_arguments) for commitment ABI & macros
- lending: drop unnecessary cast in risk_monitor utilization
- hello-world governance: votes_per_proposal via checked_div

Made-with: Cursor

+++ stellar-lend/contracts/lending/src/risk_monitor.rs
@@ -59,7 +59,7 @@ pub fn on_utilization_changed(env: &Env, total_debt: i128, debt_ceiling: i128) {
         return;
     }
 
-    let util_bps_u128 = (total_debt as i128)
+    let util_bps_u128 = total_debt
         .saturating_mul(10000)
         .saturating_div(debt_ceiling)
         .max(0) as u128;
